+++ client/app/js/task_show_dialog.jsx
@@ -4,9 +4,11 @@ import ReactMarkdown from 'react-markdown';
 import BoardActions from './actions/BoardActions';
 import BoardStore from './store/BoardStore';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
+import {Card, CardHeader, CardText} from 'material-ui/Card';
 import {BottomNavigation, BottomNavigationItem} from
   'material-ui/BottomNavigation';
 import DeleteForeverIcon from 'material-ui/svg-icons/action/delete-forever';
+import EditIcon from 'material-ui/svg-icons/editor/mode-edit';
 
 const icon_style = {
   display: "initial"
@@ -18,7 +20,8 @@ export default class TaskShowDialog extends React.Component {
     this.state = {
       task_id: null,
       name: "",
-      desc: ""
+      desc: "",
+      user: null,
     };
   }
 
@@ -31,29 +34,54 @@ export default class TaskShowDialog extends React.Component {
         task_id: task.id,
         name: task.name,
         desc: task.desc,
+        user: task.user
       });
     }
   }
 
   onClickDestroy = () => {
-    BoardActions.deleteTask(this.props.task_id);
+    BoardActions.showConfirmation(() => {
+      BoardActions.deleteTask(this.props.task_id);
+      BoardActions.closeTaskShowDialog();
+    });
+  }
+
+  onClickEdit = () => {
+    BoardActions.showTaskDialog(this.props.task_id);
   }
 
   showForm = () => {
+    var img = undefined;
+    if (this.state.user) {
+      if (this.state.user.image_path) {
+        img = this.state.user.image_path;
+      } else {
+        img = "/images/users/" + this.state.user.name +  ".png";
+      }
+    }
     return(
       <Dialog
-        title={this.state.name}
         modal={false}
         open={this.props.open}
         onRequestClose={BoardActions.closeTaskShowDialog}
         autoScrollBodyContent={true}
-      >
+      ><Card className="task">
+          <CardHeader titleStyle={{fontSize: 20}}
+            title={this.state.name}
+            avatar={img}
+          />
+          <CardText className="task-text">
+            <ReactMarkdown source={this.state.desc}/>
+          </CardText>
+        </Card>
         <br />
-        <ReactMarkdown source={this.state.desc}/>
         <BottomNavigation>
           <BottomNavigationItem label="Delete"
             icon={<DeleteForeverIcon style={icon_style} />}
             onTouchTap={this.onClickDestroy} />
+          <BottomNavigationItem label="Edit"
+            icon={<EditIcon style={icon_style} />}
+            onTouchTap={this.onClickEdit} />
         </BottomNavigation>
       </Dialog>
     );
